Rename gatsby-theme-blog-core to gatsby-theme-posts-core Fixes #616
