+++ .changeset/nice-tools-act.md
@@ -1,6 +1,6 @@
 ---
-'@undataforum/gatsby-theme-blog-core': minor
 '@undataforum/gatsby-theme-events-core': minor
+'@undataforum/gatsby-theme-posts-core': minor
 '@undataforum/gatsby-theme-profiles-core': minor
 '@undataforum/gatsby-theme-theme-ui': minor
 ---
