+++ examples/core-themes/package.json
@@ -13,7 +13,7 @@
   "dependencies": {
     "@mdx-js/mdx": "^1.6.22",
     "@mdx-js/react": "^1.6.22",
-    "@undataforum/gatsby-theme-blog-core": "*",
+    "@undataforum/gatsby-theme-posts-core": "*",
     "@undataforum/gatsby-theme-profiles-core": "*",
     "gatsby": "^2.29.2",
     "gatsby-plugin-mdx": "^1.7.1",
