+++ examples/undataforum/package.json
@@ -16,7 +16,7 @@
     "@mdx-js/react": "^1.6.22",
     "@undataforum/assets": "^0.9.0",
     "@undataforum/gatsby-theme-base": "*",
-    "@undataforum/gatsby-theme-blog-core": "*",
+    "@undataforum/gatsby-theme-posts-core": "*",
     "@undataforum/gatsby-theme-events-core": "*",
     "@undataforum/gatsby-theme-profiles-core": "*",
     "@undataforum/gatsby-theme-theme-ui": "*",
