+++ packages/gatsby-theme-posts-core/gatsby-config.js
