+++ packages/gatsby-theme-posts-core/gatsby-node.js
