+++ packages/gatsby-theme-posts-core/index.js
