+++ packages/gatsby-theme-posts-core/src/components/post-page.js
