+++ packages/gatsby-theme-posts-core/src/components/posts-page.js
