+++ packages/gatsby-theme-posts-core/src/post-fragment.js
