+++ packages/gatsby-theme-posts-core/src/templates/post-query.js
