+++ packages/gatsby-theme-posts-core/src/templates/posts-query.js
