+++ packages/gatsby-theme-posts-core/theme-options.js
